[tmpnet] Update URI and StakingAddress usage in support of kube
Previously, Node.URI and Node.StakingAddress fields contained values
for a local process and were always locally accessible. Nodes running
in kubernetes may be running remotely, though, and only be accessible
if e.g. a local port is forwarded to the port of a node's pod. To
ensure compatibility with forwarded URIs and StakingAddresses, callers
that aren't guaranteed to be running locally to a node are updated to
access these values via functions that can ensure a remote endpoint is
forwarded if the target node isn't running locally.