This commit removes the toEngine channel pattern from all VMs and adds an additional API:
```
SubscribeToEvents(ctx context.Context, pChainHeight uint64) (Message, uint64)
```

Which returns when an event (pending transactions or state sync done Message) is sent by the VM. The SubscribeToEvents receives additionally a P-chain height, which is used in the Warp precompile.

Unlike the toEngine pattern which only notifies the consumer a single time until a block is built, successive calls to `SubscribeToEvents` will return that transactions are pending as long as the mem-pool is not empty and a block can be built, otherwise the calls block.

The API change is needed to support consensus protocols which facilitate censorship resistance as part of their protocol and therefore need information about whether a block should be built or not on demand, in contrast to the curren toEngine pattern which notifies the consensus when new transactions arrive only once, and then block building may even fail. Having such an API on the VM side (and not implementing the logic in consensus) forces the VM to handle any side effects of block building (such as re-introducing transactions into the mempool).

Signed-off-by: Yacov Manevich <yacov.manevich@avalabs.org>